requirements.txt
torch
torchvision
matplotlib
seaborn
numpy
scikit-learn
PyYAML
opencv-python
torchviz
torchview
